fix: move JMT stats from periodic stats output to separate command
